[Task dialog] - Added asynchronous validation for tasks
