+++ ui/src/components/Tasks/Characterisation.js
@@ -4,6 +4,7 @@ import { connect } from 'react-redux';
 import { reduxForm, formValueSelector } from 'redux-form';
 import { Modal, Button, Form, Row, Col, ButtonToolbar } from 'react-bootstrap';
 import { DraggableModal } from '../DraggableModal';
+import asyncValidate from './asyncValidate';
 import validate from './validate';
 import warn from './warning';
 
@@ -466,6 +467,7 @@ class Characterisation extends React.Component {
 
 const CharacterisationForm = reduxForm({
   form: 'characterisation',
+  asyncValidate,
   validate,
   warn,
 })(Characterisation);
